# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovaRoute API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/student/recommendations": {
      "get": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Recommended",
        "operationId": "recommended_api_student_recommendations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1,
              "minimum": 0,
              "default": 0.8,
              "title": "Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recommendations/{student_id}": {
      "get": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Recommended Id",
        "operationId": "recommended_id_api_recommendations__student_id__get",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1,
              "minimum": 0,
              "default": 0.8,
              "title": "Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/student/saved": {
      "get": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Saved",
        "operationId": "saved_api_student_saved_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SavedOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/student/saved/{identifier}": {
      "post": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Save",
        "operationId": "save_api_student_saved__identifier__post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Note"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Save Note",
        "operationId": "save_note_api_student_saved__identifier__patch",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Note"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Unsave",
        "operationId": "unsave_api_student_saved__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications": {
      "get": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Inbox",
        "operationId": "inbox_api_notifications_get",
        "parameters": [
          {
            "name": "read",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(all|unread|read)$",
              "default": "all",
              "title": "Read"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NotificationOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/{identifier}": {
      "patch": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Read Notification",
        "operationId": "read_notification_api_notifications__identifier__patch",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recommendations, saved listings, notifications"
        ],
        "summary": "Dismiss",
        "operationId": "dismiss_api_notifications__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "ListingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "employer_id": {
            "type": "integer",
            "title": "Employer Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "work_mode": {
            "type": "string",
            "title": "Work Mode"
          },
          "duration": {
            "type": "string",
            "title": "Duration"
          },
          "stipend_min": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stipend Min"
          },
          "stipend_max": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stipend Max"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "admin_hidden": {
            "type": "boolean",
            "title": "Admin Hidden",
            "default": false
          },
          "content_version": {
            "type": "integer",
            "title": "Content Version"
          },
          "provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Provenance"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyOut"
          },
          "moderation": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderation"
          },
          "moderation_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderation Model"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "employer_id",
          "title",
          "description",
          "required_skills",
          "location",
          "work_mode",
          "duration",
          "stipend_min",
          "stipend_max",
          "currency",
          "deadline",
          "status",
          "content_version",
          "provenance",
          "company",
          "created_at",
          "updated_at"
        ],
        "title": "ListingOut"
      },
...
      "MatchOut": {
        "properties": {
          "internship": {
            "$ref": "#/components/schemas/ListingOut"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity"
          },
          "percentage": {
            "type": "number",
            "title": "Percentage"
          },
          "matched_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Skills"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "score_help": {
            "type": "string",
            "title": "Score Help"
          }
        },
        "type": "object",
        "required": [
          "internship",
          "similarity",
          "percentage",
          "matched_skills",
          "missing_skills",
          "explanation",
          "score_help"
        ],
        "title": "MatchOut"
      },
      "Note": {
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Note"
      },
      "NotificationEdit": {
        "properties": {
          "read": {
            "type": "boolean",
            "title": "Read"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "read"
        ],
        "title": "NotificationEdit"
      },
      "NotificationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "recipient_id": {
            "type": "integer",
            "title": "Recipient Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "related": {
            "$ref": "#/components/schemas/Related"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Feedback"
              },
              {
                "type": "null"
              }
            ]
          },
          "feedback_state": {
            "type": "string",
            "title": "Feedback State"
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "dismissed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dismissed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "email_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email State"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recipient_id",
          "type",
          "title",
          "body",
          "related",
          "feedback",
          "feedback_state",
          "read_at",
          "dismissed_at",
          "created_at"
        ],
        "title": "NotificationOut"
      },
      "OK": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "OK"
      },
...
      "Page_NotificationOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[NotificationOut]"
      },
...
      "Page_SavedOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SavedOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[SavedOut]"
      },
...
      "RecommendationsOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchOut"
            },
            "type": "array",
            "title": "Items"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "strictly_greater": {
            "type": "boolean",
            "title": "Strictly Greater"
          }
        },
        "type": "object",
        "required": [
          "items",
          "threshold",
          "strictly_greater"
        ],
        "title": "RecommendationsOut"
      },
...
      "SavedOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "internship_id": {
            "type": "integer",
            "title": "Internship Id"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available"
          },
          "internship": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListingOut"
              },
              {
                "$ref": "#/components/schemas/UnavailableListing"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internship"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "internship_id",
          "note",
          "created_at",
          "updated_at"
        ],
        "title": "SavedOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}